+++ Frameio.NET.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28803.156
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frameio.NET", "src\Frameio.NET\Frameio.NET.csproj", "{4CBDD5AE-D7A1-40DD-B629-46CCA9C889D2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frameio.NET.Tests", "tests\Frameio.NET.Tests\Frameio.NET.Tests.csproj", "{A3FC474B-332C-4C3D-A011-9BACB05EC9B1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4CBDD5AE-D7A1-40DD-B629-46CCA9C889D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CBDD5AE-D7A1-40DD-B629-46CCA9C889D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CBDD5AE-D7A1-40DD-B629-46CCA9C889D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CBDD5AE-D7A1-40DD-B629-46CCA9C889D2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A3FC474B-332C-4C3D-A011-9BACB05EC9B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A3FC474B-332C-4C3D-A011-9BACB05EC9B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A3FC474B-332C-4C3D-A011-9BACB05EC9B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A3FC474B-332C-4C3D-A011-9BACB05EC9B1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CA78807C-42EE-47F6-A1E5-CF0A91C91DA5}
+	EndGlobalSection
+EndGlobal
