--- src/FrameIoNet/Frameio.NET/Models/ErrorResponse.cs
@@ -1,11 +1,0 @@
-﻿namespace Frameio.NET.Models {
-    public class ErrorResponse
-    {
-        public int Code { get; set; }
-
-        public Error[] Errors { get; set; }
-
-        public string Message { get; set; }
-
-    }
-}
