--- src/FrameIoNet/Frameio.NET/Models/PagedResult.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Frameio.NET.Models
-{
-    public class PagedResult<T>
-    {
-        public Paging Paging { get; set; }
-
-        public IEnumerable<T> Results { get; set; }
-
-    }
-}
