+++ src/Frameio.NET/Models/Error.cs
@@ -1,0 +1,14 @@
+﻿namespace Frameio.NET.Models
+{
+    public class Error {
+
+        public string Code { get; set; }
+
+        public string Detail { get; set; }
+
+        public int Status { get; set; }
+
+        public string Title { get; set; }
+
+    }
+}
