+++ src/Frameio.NET/Models/ErrorResponse.cs
@@ -1,0 +1,11 @@
+﻿namespace Frameio.NET.Models {
+    public class ErrorResponse
+    {
+        public int Code { get; set; }
+
+        public Error[] Errors { get; set; }
+
+        public string Message { get; set; }
+
+    }
+}
