+++ src/Frameio.NET/Models/User.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+
+namespace Frameio.NET.Models
+{
+    public class User
+    {
+        public string Id { get; set; }
+
+        [JsonProperty("account_id")]
+        public string AccountId { get; set; }
+
+        public string Email { get; set; }
+
+        public string Name { get; set; }
+
+    }
+}
